add query for geotags & Limit for queries
Now the dataLoader function getJournalEntries accepts a
"number of entries" argument and will limit the database query
with that number. So we can ask for only 10 entries, or only 20 entries.

Also now there is a new dataLoader function getGeotaggedEntries,
which functions exactly like getJournalEntries, except
it only returns entries with a non-null "lat" value. So
if an entry has a latitude number, it'll be returned.